+++ index.html
@@ -3,25 +3,21 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="style/style.css" />
-    <link rel="stylesheet" href="style/main.css">
+    <link rel="stylesheet" href="style/main/style.css" />
+    <link rel="stylesheet" href="style/main/main.css">
     <link rel="preconnect" href="https://fonts.googleapis.com" />
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
     <link
       href="https://fonts.googleapis.com/css2?family=Bebas+Neue&family=Black+Han+Sans&family=Do+Hyeon&family=Nanum+Pen+Script&family=Roboto+Slab:wght@500&display=swap"
       rel="stylesheet"
     />   
-   
-    
-    <script type="module" src="./src/main.js"></script>
-    <script type="module" src="./src/movie.js"></script>
-    <script type="module" src="./src/rediretion.js"></script>
-    
+    <script type="module" src="./src/main/main.js"></script>
+    <script type="module" src="./src/main/movie.js"></script>
+    <script type="module" src="./src/main/rediretion.js"></script>
 <script
 src="https://kit.fontawesome.com/1f700c87c6.js"
 crossorigin="anonymous"
 ></script>
-
 <script
 src="https://kit.fontawesome.com/1f700c87c6.js"
 crossorigin="anonymous"
