+++ src/detail/sub.js
@@ -1,9 +1,9 @@
-import updateReviewList from "./statelist.js";
+import updateReviewList from "../review/statelist.js";
 
 const urlParams = new URLSearchParams(window.location.search);
 const itemId = urlParams.get("id");
 console.log("Item ID:", itemId);
-const API_KEY = api.key;
+const API_KEY = "api_key=2f7ff395e001967bcd029e4d663de74c";
 const BASE_URL = "https://api.themoviedb.org/3/movie/";
 const FULL_URL = `${BASE_URL + itemId}?language=en-US&${API_KEY}`;
 
@@ -21,18 +21,22 @@ const showMovieDetails = (movie) => {
   const IMG_URL = "https://image.tmdb.org/t/p/w500";
   const POSTER_SECTION = document.getElementById("poster-section");
   const INFO_SECTION = document.getElementById("info-section");
+  const REVIEW_SECTION = document.getElementById("review-section");
 
   POSTER_SECTION.innerHTML = `
         <img src=${IMG_URL + movie.poster_path} />
   `;
 
   INFO_SECTION.innerHTML = `
-        <h1>${movie.title}</h1>
-        <p class="releas-date">${movie.release_date}</p>
-        <p class="vote-star">
-          <img src="../assets/main/star.svg" />
-          ${Math.ceil((movie.vote_average / 2) * 10) / 10}
-        </p>
+        <div class="title-box">
+          <h1>${movie.title}
+            <p class="releas-date">${movie.release_date}</p>
+          </h1>
+          <p class="vote-star">
+            <img src="../assets/main/star.svg" />
+            ${Math.ceil((movie.vote_average / 2) * 10) / 10}
+          </p>
+        </div>
         <div class="details-box">
           <div class="film-box">
             <img class="film-bar" src="../assets/main/film_bar_gray.png">
@@ -47,9 +51,12 @@ const showMovieDetails = (movie) => {
             </a>
           </div>
         </div>
-        <div id="review-box">
-        </div>
-  `;
+        `;
+
+  // REVIEW_SECTION.innerHTML = `
+  //       <div id="review-box">
+  //       </div>
+  // `;
 
   updateReviewList(movie.id);
 };
