--- style/sub_2.css
@@ -1,137 +1,0 @@
-* {
-  margin: 0;
-  padding: 0;
-}
-
-body {
-  background-color: #000;
-  color: #fff;
-}
-
-nav {
-  position: relative;
-  height: 25px;
-  padding: 18px 35px;
-}
-
-nav img {
-  position: absolute;
-  top: 50%;
-  left: 35px;
-  width: 175px;
-  transform: translate(0, -50%);
-}
-
-nav button {
-  position: absolute;
-  top: 50%;
-  right: 35px;
-  width: 100px;
-  height: 35px;
-  background-color: #02a7f7;
-  border: none;
-  border-radius: 5px;
-  color: #fff;
-  font-size: 1rem;
-  transform: translate(0, -50%);
-}
-
-#movie-article {
-  display: grid;
-  grid-template-columns: 40% 60%;
-}
-
-#poster-section {
-  margin: 20px 0;
-  text-align: center;
-}
-
-#poster-section img {
-  height: 85vh;
-}
-
-#info-section {
-  margin: 20px 0;
-  padding-right: 40px;
-  text-align: left;
-}
-
-#info-section h1 {
-  display: inline-block;
-  font-size: 3cqmax;
-  padding: 10px 0 20px 0;
-}
-
-.releas-date {
-  display: inline-block;
-  font-size: 0.8rem;
-  padding-left: 7px;
-  color: #c1c1c1;
-}
-
-.vote-star {
-  padding-top: 30px;
-  float: right;
-  font-size: 2rem;
-  text-align: center;
-}
-
-.vote-star img {
-  width: 30px;
-}
-
-.details-box {
-  display: flex;
-  margin-bottom: 50px;
-}
-
-.film-box {
-  display: flex;
-  /* float: left; */
-  justify-content: space-between;
-  width: 60%;
-  background-color: #000;
-  border-radius: 10px;
-  box-shadow: 4px 4px 30px 2px #02a5f7bc;
-  align-items: center;
-  padding: 0 15px;
-  margin-right: 35px;
-}
-
-.back-porster {
-  height: 25vh;
-  padding: 30px;
-}
-
-.film-bar {
-  height: 30vh;
-}
-
-.info-box {
-  padding: 15px;
-  border-radius: 10px;
-}
-
-.tag-line {
-  font-size: 1.5rem;
-  font-style: italic;
-  text-shadow: 3px 3px 10px #02a7f7;
-  padding-bottom: 25px;
-}
-
-.plot {
-  font-size: 13px;
-  color: #9e9e9e;
-  padding-bottom: 20px;
-}
-
-.homepage-btn img {
-  /* float: right; */
-  height: 40px;
-}
-
-#review-box {
-  height: 30vh;
-  border: 2px solid #02a5f783;
-  border-radius: 10px;
-}
