--- sub/index_copy.html
@@ -1,54 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="../style/style.css" />
-    <link rel="preconnect" href="https://fonts.googleapis.com" />
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
-    <link
-      href="https://fonts.googleapis.com/css2?family=Bebas+Neue&family=Black+Han+Sans&family=Do+Hyeon&family=Nanum+Pen+Script&family=Roboto+Slab:wght@500&display=swap"
-      rel="stylesheet"
-    />
-    <script type="module" src="../src/movie_copy.js"></script>
-    <title>search MOVIE</title>
-  </head>
-  <body>
-    <!--header-->
-    <header>
-      <div class="navbar">
-        <div>
-          <img src="../assets/main/1. 로고.png" class="logo" />
-          <a class="text">PROJECT.1</a>
-        </div>
-        <div class="title">
-          <h1>SPARTA MOVIE</h1>
-        </div>
-        <div>
-          <img src="../assets/main/2. 아이콘.png" class="logo2" />
-        </div>
-      </div>
-    </header>
-    <!--search-->
-    <nav>
-      <form class="section1" id="form">
-        <input id="search" class="input" type="text" placeholder="        Search" autofocus />
-      </form>
-
-      <!--movie list-->
-      <main id="main">
-        <div class="movie">
-          <img class="card" src="" alt="Image" />
-          <div class="movieinfo">
-            <h3>Movie Title</h3>
-            <span class="green">9.8</span>
-          </div>
-          <div class="overview">
-            <h3>Overview</h3>
-            what's your position
-          </div>
-        </div>
-      </main>
-    </nav>
-  </body>
-</html>
