+++ sub/login.html
@@ -10,9 +10,9 @@
       crossorigin="anonymous"
     />
     <link rel="stylesheet" href="../style/reset.css" />
-    <link rel="stylesheet" href="../style/login.css" />
+    <link rel="stylesheet" href="../style/login/login.css" />
 
-    <script type="module" src="../src/login.js"></script>
+    <script type="module" src="../src/login/login.js"></script>
     <title>Document</title>
   </head>
   <body>
