+++ sub/profile.html
@@ -4,8 +4,8 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
 
-    <link rel="stylesheet" href="../style/test.css" />
-    <script type="module" src="../src/profile.js"></script>
+    <link rel="stylesheet" href="../style/profile/profile.css" />
+    <script type="module" src="../src/profile/profile.js"></script>
     <title>Document</title>
   </head>
   <body>
