+++ sub/signUp.html
@@ -3,8 +3,8 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="../style/login.css" />
-    <script type="module" src="../src/signup.js"></script>
+    <link rel="stylesheet" href="../style/login/login.css" />
+    <script type="module" src="../src/login/signup.js"></script>
 
     <title>회원가입 페이지</title>
   </head>
